lenet/requirements.txt
# Neural Network Model Dependencies
pytorch
torchmetrics
torchvision

# Core Python Libraries
numpy

# Visualization Tools
tensorboard
matplotlib
mlflow

# Utility Tools
tqdm
pyyaml